Separa os dados do jogador em diversos arquivos para modularização
